+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Feature
 
+- Support the use of arguments on simple fields.
 - Possibility to use an alternative backendResourceName in resource options
 
 ## 2020-02-17 / 0.7.1